Ships movement and limitations + Keypresses
Movement
I setup ship movement as a boolean so that when 'a' or 'd' are pressed, the boolean is set to true and the ships movement will go continuously by a factor of 1.5 pixels. When the key is released, the boolean is set to false and the movement of the ship has stopped.

Limitations
I set it so that that the ship is limited to the boundaries of the window.

Key presses
a - left 
d - right
q - quit game